refactor: extract vaadin-virtual-list-mixin and split tests
