+++ dev/virtual-list.html
@@ -1,0 +1,30 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Virtual list</title>
+    <script type="module" src="./common.js"></script>
+
+    <script type="module">
+      import '@vaadin/virtual-list';
+
+      const items = Array.from({ length: 100000 }).map((_, i) => {
+        return { label: `Item ${i}` };
+      });
+
+      const renderer = (root, _, { item }) => {
+        root.innerHTML = `<div>${item.label}</div>`;
+      };
+
+      const list = document.querySelector('vaadin-virtual-list');
+      list.items = items;
+      list.renderer = renderer;
+    </script>
+  </head>
+
+  <body>
+    <vaadin-virtual-list style="height: 400px"></vaadin-virtual-list>
+  </body>
+</html>
