+++ packages/virtual-list/package.json
@@ -37,6 +37,7 @@
     "web-component"
   ],
   "dependencies": {
+    "@open-wc/dedupe-mixin": "^1.3.0",
     "@polymer/polymer": "^3.0.0",
     "@vaadin/component-base": "24.3.0-alpha10",
     "@vaadin/lit-renderer": "24.3.0-alpha10",
