+++ packages/virtual-list/src/vaadin-virtual-list.d.ts
@@ -3,22 +3,16 @@
  * Copyright (c) 2021 - 2023 Vaadin Ltd.
  * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
  */
-import { ControllerMixin } from '@vaadin/component-base/src/controller-mixin.js';
 import { ElementMixin } from '@vaadin/component-base/src/element-mixin.js';
 import { ThemableMixin } from '@vaadin/vaadin-themable-mixin/vaadin-themable-mixin.js';
+import type {
+  VirtualListDefaultItem,
+  VirtualListItemModel,
+  VirtualListMixinClass,
+  VirtualListRenderer,
+} from './vaadin-virtual-list-mixin.js';
 
-export type VirtualListDefaultItem = any;
-
-export interface VirtualListItemModel<TItem> {
-  index: number;
-  item: TItem;
-}
-
-export type VirtualListRenderer<TItem> = (
-  root: HTMLElement,
-  virtualList: VirtualList<TItem>,
-  model: VirtualListItemModel<TItem>,
-) => void;
+export { VirtualListDefaultItem, VirtualListItemModel, VirtualListRenderer };
 
 /**
  * `<vaadin-virtual-list>` is a Web Component for displaying a virtual/infinite list of items.
@@ -46,51 +40,11 @@ export type VirtualListRenderer<TItem> = (
  * @extends HTMLElement
  * @mixes ElementMixin
  * @mixes ThemableMixin
+ * @mixes VirtualListMixin
  */
-declare class VirtualList<TItem = VirtualListDefaultItem> extends ElementMixin(
-  ControllerMixin(ThemableMixin(HTMLElement)),
-) {
-  /**
-   * Gets the index of the first visible item in the viewport.
-   */
-  readonly firstVisibleIndex: number;
+declare class VirtualList<TItem = VirtualListDefaultItem> extends ThemableMixin(ElementMixin(HTMLElement)) {}
 
-  /**
-   * Gets the index of the last visible item in the viewport.
-   */
-  readonly lastVisibleIndex: number;
-
-  /**
-   * Custom function for rendering the content of every item.
-   * Receives three arguments:
-   *
-   * - `root` The render target element representing one item at a time.
-   * - `virtualList` The reference to the `<vaadin-virtual-list>` element.
-   * - `model` The object with the properties related with the rendered
-   *   item, contains:
-   *   - `model.index` The index of the rendered item.
-   *   - `model.item` The item.
-   */
-  renderer: VirtualListRenderer<TItem> | undefined;
-
-  /**
-   * An array containing items determining how many instances to render.
-   */
-  items: TItem[] | undefined;
-
-  /**
-   * Scroll to a specific index in the virtual list.
-   */
-  scrollToIndex(index: number): void;
-
-  /**
-   * Requests an update for the content of the rows.
-   * While performing the update, it invokes the renderer passed in the `renderer` property for each visible row.
-   *
-   * It is not guaranteed that the update happens immediately (synchronously) after it is requested.
-   */
-  requestContentUpdate(): void;
-}
+interface VirtualList<TItem = VirtualListDefaultItem> extends VirtualListMixinClass<TItem> {}
 
 declare global {
   interface HTMLElementTagNameMap {
