+++ packages/virtual-list/test/lit-polymer.test.js
@@ -1,0 +1,2 @@
+import '../vaadin-virtual-list.js';
+import './lit.common.js';
