+++ packages/virtual-list/test/lit-renderer-directives-polymer.test.js
@@ -1,0 +1,2 @@
+import '../vaadin-virtual-list.js';
+import './lit-renderer-directives.common.js';
