+++ packages/virtual-list/test/lit.common.js
@@ -1,6 +1,5 @@
 import { expect } from '@esm-bundle/chai';
 import { fixtureSync } from '@vaadin/testing-helpers';
-import '../vaadin-virtual-list.js';
 import { html, render } from 'lit';
 
 describe('lit', () => {
