+++ packages/virtual-list/test/virtual-list-polymer.test.js
@@ -1,0 +1,2 @@
+import '../vaadin-virtual-list.js';
+import './virtual-list.common.js';
