+++ packages/virtual-list/test/virtual-list.common.js
@@ -1,6 +1,5 @@
 import { expect } from '@esm-bundle/chai';
 import { fixtureSync, nextFrame } from '@vaadin/testing-helpers';
-import '../vaadin-virtual-list.js';
 
 describe('virtual-list', () => {
   let list;
